decomposed + vernam
