+++ errors.py
@@ -1,0 +1,8 @@
+class ArgsNumError(Exception):
+    def __init__(self, text):
+        self.txt = text
+
+
+class TaskTypeError(Exception):
+    def __init__(self, text):
+        self.txt = text
