+++ output_ve_de.txt
